#134 refactor userdataprovider into plugin-api.events
