+++ fullstop-plugin-api/pom.xml
@@ -40,6 +40,10 @@
             <artifactId>json-path</artifactId>
         </dependency>
         <dependency>
+        	<groupId>org.yaml</groupId>
+        	<artifactId>snakeyaml</artifactId>
+        </dependency>
+        <dependency>
             <groupId>org.springframework.plugin</groupId>
             <artifactId>spring-plugin-core</artifactId>
         </dependency>
